CMPLRLLVM-32942: Update Intel argument alignment for opaque pointers (#2180)
This change set generalize the pointer type checks for the Intel
argument aligment in order to support opaque and types pointers.
The changes removes the check for i8* type, and only checks
if the pointer is casted to an i64. The argument aligment
analysis process checks if the pointer is aligned by 8 bytes.