[DPCPPKernelTransforms] Add dependency on BuiltinLibInfoAnalysis (#2205)
BuiltinImport, DPCPPEqualizer, GroupBuiltin and ResolveSubGroupWICall
passes depends on BuiltinLibInfoAnalysis.
Move RuntimeService.h to include folder and use it as member variable
in BuiltinLibInfo.

We can't load BuiltinModules in BuiltinLibInfoAnalysis because it delays
load time from clCreateContext to clBuildProgram and thus causes ~45ms
compile time regression. So we pass BuiltinModules as parameter to it.