Stop generating debug info for cpu-dispatch resolver when ifunc disabled (#2244)
As reported in CMPLRLLVM-33883, when ifuncs are disabled in linux mode
and debug information is enabled, the debug information on the
'resolver' mixed with the musttail calls to the variants causes a
verification failure.

The failure is because we require tailedcalled-calls to have debug info
if the containing function also has debug info.  However, it doesn't
really make sense for the resolver to have debug info, as it doesn't
have any source code locations to be associated with.  Instead, this
patch suppresses the debug information for this situation.

Note that this is only a problem with the -disable-cpudispatch-ifuncs
flag, which is xmain specific, as the ifunc doesn't get debug
information, and the windows multiversion calls don't seem to have this
problem at all (they don't generate debug info for the resolver
already).