[VPlan][VPEntities] Don't assume that conditional last private's exit PHI is in loop's latch. (#2199)
While lowering conditional last private entities we emit a new
PHI in loop header to track the last index for which condition
is true. We assume that the private's exit PHI is always in the
loop latch - in reality it could be in any BB dominating latch.
This assumption leads to a buggy PHI node after VPEntities
transform.

The exit PHI can be in any BB that dominates loop latch. The
current assumption leads to generation of invalid loop header
PHIs like -

%idx.phi = phi [ -1, %preheader ], [ %idx.phi.nxt, %non.latch.bb ]

This patch removes this assumption while lowering conditional
last private VPEntities.

JIRA: CMPLRLLVM-33891