[SYCL] cherry pick xpti  accessor events (#2225)
It seems that some new changes to the llvm-test-suite are failing with xmain-cand. This is because the test suite is referencing some enum values that haven't been brought over into xmain yet.

Here I am cherry picking the commits from intel/llvm#5249 to xmain. This makes xmain (and xmain-cand) pass the currently failing xpti tests.