CMPLRLLVM-33537: Disable partial load hoisting when HIR AVX512 is enabled. (#2169)
This creates an extra phi, as per https://reviews.llvm.org/D113289. The
extra loop-carried value will cause HIR to perform more transformations, which
will have a negative effect on vectorization.

Applies only when compiling for AVX512, as it looks like the load hoisting helps one of the coremark benchmarks.