[Dynamic LoopOpt] Implement support in Old Pass Manager (#2237)
Dynamic loopopt support in the new PM will be different (and much simpler) because
the Function/Loop PM is itself a pass there. So, it's easy to wrap that with a
thin helper class that will return early and not apply the whole set of passes
at once.

In the Old PM, we customize the Pass itself and assign a "Limiter" to it. We
also add aditional checks inside the Function/Loop pass managers' to not run
the pass if it's limited and function's attributes indicate that the pass needs
to be skipped on that function.

Note that we enable loopopt pipeline on a given function via a dedicated attribute,
i.e. not as part of TTI-like analysis as that wouldn't be accessible from that place.

* Fix tests after rebase

Co-authored-by: Andrei Elovikov <andrei.elovikov@intel.com>