[NFC] CMPLRLLVM-33397: Fix for memory leak in DTrans due to Dominator Tree (#2280)
This change set removes the recomputation the dominator tree
in DTrans analysis. The DTrans analysis invokes the block
frequenzy analysis, which calls branch probability and loop
info analyses. Loop info calls the dominator tree as
required transitive, which means that the information of
the analysis will be alive while the pass that uses it is
alive. If the pointer to the dominator tree is detached
to compute a new dominator tree, then there could be a memory
leak.