[VPlan][HIR] Remove temporary workaround that prohibits masked fabs vectorization. (#2277)
In order to fix the regression reported in CMPLRLLVM-11468
we added a temporary bailout mechanism in VPlan HIR vector
CG to skip loops containing masked fabs intrinsic. This was
meant as a temporary placeholder until AZB feature was
implemented and tuned in VPlan. This patch removes this
workaround as it was verified locally that original
regression has been addressed in latest xmain and performance
improves upon vectorization of problematic loop.

JIRA : CMPLRLLVM-33628