[SYCL] Fix kernel program cache for multiple devices and refactor some unit tests (#2315)
When multiple devices are used in the same context, the problem of piProgram double free is arised.
In program_manager::build() created program is shared with multiple devices and is cached in KernelProgramCache as many times as there are devices. When cache is destructed, the same program will be released for every device, but piProgramRetain was called only for one device, which leads to double free. Adding piProgramRetain for each device solves this problem.
Deleting kernel from cache after its release prevents double free in the case a kernel was created for a program that shared with multiple devices.
Unittest with emulating multiple gpu is added.
Signed-off-by: mdimakov <maxim.dimakov@intel.com>

This is a cherry-pick of PR#5017 from intel/llvm to xmain.