[CMPLRLLVM-34115] [DTrans - OpaquePtr] Extend DTransSafetyAnalyzer for 538.imagick (#2394)
Extend the DTransSafetyAnalyzer to handle a case where a pointer that is
newly allocated is used as an 'i8*' type and stored to a 'double*' structure
field when all the uses as the i8* are safe. Also, update a location where
a safety flag is set to make sure it goes through the path that reports
the trace messages.