CMPLRLLVM-33870: Libfuncs for 520.omnetpp and new pass manager (#2336)
The new pass manager may apply different optimizations that may
affect the functions available in the module (e.g. inlining,
call delete, etc.). There is a chance that we may have libfuncs
with IR in the module and may affect the whole program analysis.
This is one of those cases. The changes add some extra libfuncs
with IR that aren't reflected in the legacy pass manager.