CMPLRLLVM-34170: Ensure that setting readonly and writeonly among libFuncs do not conflict. (#2452)
* CMPLRLLVM-34170: Ensure that setting readonly and writeonly among libFuncs
do not conflict.

We have an Intel customization in BuildLibCalls.cpp that sometimes sets a
libFunc to readonly.  The community wants to set some of these libFuncs to
writeonly, but if they are readonly the readonly should be changed to
readnone.  We can accomplish this by first setting these LibFuncs as
writeonly and only after that setting them to readonly. This will cause
them to be set to readnone.

I ran the LIT tests on this by hand, but will also test this change on an
xmain change set to make sure there are no surprise regressions.

* Take Axel's suggestion