CMPLRLLVM-34228: Fix DTransBadCastingAnalyzer::isPotentialBitCastOfAllocStore (#2470)
Klocwork reported another 41 issues, and in this case 2 of them (one was a
duplicate of the other) indicated a real issue. On closer inspection, I noted
that the function DTransBadCastingAnalyzer::isPotentialBitCastOfAllocStore
was incorrectly implemented, as it did not match the description in the
comment above the function. I have fixed it and added LIT tests.  This is not
an issue in the opaque pointer version of the implementation, as there are no
bitcasts present for which to test.