[DTrans - OpaquePtr] Improve trace messages for DTrans type metadata reader test pass (#2473)
The -dtrans-typemetadatareader pass is used to report information about the
metadata used for DTrans type information for testing purposes. This change
cleans up the report to print the relevant information on a single line. Also,
provide some control flags to allow limiting the information to simplify
finding specific kinds of issues.