[sycl-post-link] Add handling for zero-initialized array constants (#5321) (#2466)
The problem is the following, when the `getOperand` method is invoked for a
`zeroinitializer` array: `C->getOperand(i)`, the assert from `User.h` was triggered:

```
    assert(i < NumUserOperands && "getOperand() out of range!");
```

This is because `zeroinitializer` has no operands. The problem occurs only when
the specialization constant was not initialized directly in the point of definition.

When a specialization constant is defined in the IR as `zeroinitializer`, its default
value is just an array of zeros of the arbitrary size. No recursive calls of the
`collectCompositeElementsDefaultValuesRecursive` function is required in the case.

Also, to make processing constants with different types more robust, dynamic
casting of the constant itself (`dyn_cast<ConstantType>(C)`) should be used
instead of casting the constant's type since the type can be inconsistent with
the constant behavior, for example not every constant with the type equals to
`ArrayType` has any components (the `zeroinitializer` has not).

This a cherry-pick of intel/llvm PR#5321.