Disable auto memory with x86 windows build (#210)
This is to avoid out of memory resource error on x86 windows. And also updated
comments in cl_config.in and disabled tests for auto memory with x86 windows
build.
Replaced MessageBox for showing error message with errs() as the former was
not good for user experience and automatic testing.
Removed useless Logger in backend/ocl_cpu_backend folder.