CMPLRLLVM-34301: Fix for regression in 505.mcf after updating argument alignment in opaque pointers (#2638)
The update for argument aligment introduced a check for GEP
that broke the check for some candidates to apply the
transformation. These candidates were allocated using alloca
instructions rather than calloc since they are arrays. This
change set basically makes sure that collecting from a
GEP that is an array returns a pointer type.

Also, this change set fixes the traversal for checking the
store instructions. Originally, this traversal was
implemented as a recursion, which could produce a false
'false' if the store instruction wasn't collected from a
path and all the users were verified. It now uses a queue
to make sure all possible paths were covered to collect
the store instruction.