cmplrllvm-34811: Fix opaque pointer problem by remove unnecessary change (#2682)
The opaque pointer problem is due to the code added for firstprivate variable
with constant type.

That part of code was added for cmplrllvm-27130. Which is for change of

commit 1af5bd54a8273398d33c840b8a26852ff3be73f7

[OPENMP] Target region: emit const firstprivates as globals with constant
memory.

If the variable with the constant non-scalar type is firstprivatized in
the target region, the local copy is created with data copying.
Instead, we allocate the copy in the constant memory and avoid extra
copying in the outlined target regions. This global copy is used in the
target regions without loss of performance.

However, the whole change has been removed with
 https://reviews.llvm.org/D105375

So this change is no longer needed and is therefore removed.