CMPLRLLVM-34810: Support IP cloning in metadata inlining report (#2666)
In addition, this fixes a few bugs in the metadata inlining report I found
in handling of dead static functions. This required updating some of the LIT
tests like MDInlineReport_transform01.ll, intel-ipo-dead-arg-04.ll, and
intel-ipo-dead-arg-04-opq.ll that had hard-coded results. I extended LIT tests
InlineReport86.ll and intel-cmplrllvm-29570.ll to include testing of the
metadata inlining report and added InlineReport104.ll to catch the case of
functions that call functions to ensure taht metadata on callsites was
maintained when a function is cloned.