CMPLRLLVM-34856: Add missing DELETE entries in classic inlining report (#2717)
While inlining a call, the inliner can eliminate callsites in the copied in the callee via on-the-fly dead code elimination. Such callsites should be reported as DELETED in the inlining report. They were reported in the metdata inlining report, but not in the classic inlining report. Also, when replaing the active
inline callsite, the callback for the inlined callsite should be removed.