[VPlan][OptReport] Provide centralized OptReportStatsTracker for each VPLoop in VPlan CFG. (#2331) (#2563)
Currently OptReportStatsTracker is isolated to codegen and we
implicitly report all collected statistics for the main vector
loop. This leads to incorrect report when we have vectorized
peel/remainder loops since the statistics should be collected
and reported separately for each outgoing loop. This patch
addresses this issue by creating a central storage for
OptReportStatsTracker in VPExternalValues. The tracker is
identified by the VPLoop it is associated with.

JIRA: CMPLRLLVM-28721