cmplrllvm-33708: set RemovedReference. (#2546)
Problem is that type of "pt" and "pt->map.ptr.temp" are not same when remap
pt with pt->map.ptr.temp where call to EmitLoadOfReferenceLValue. However,
since it is just remap, the EmitLoadOfReferenceLValue should not be called.

To fix this just set RemovedReference, so that EmitLoadOfReferenceLValue
can be skipped.