[VecClone] Refactor linear argument handling (#2671)
This patch cleans up the code that does linear argument processing.
This code is responsible for generating a stride calculation and then
replacing the arg user with this stride. This code follows the same
approach as vector arg handling where we don't try to specialize
handling between -O0 input and above. We just simply generate the
stride instructions, insert them before the arg user, and update
the arg user operands.