CMPLRLLVM-34110: Update IPO Prefetch to support opaque pointers (#2748)
This change set updates the IPO prefetch to support opaque
pointers. It traces the users of the an argument and collect
the GEPs, from there we can catch if the type is a pointer
to a structure.

It also the IPO utils function matcher class. Rather than
count which arguments ars floating points and double, it
checks if the total of arguments that are integer and
pointers is the same as the total number of arguments.