[OpenMP]Propagate the ElementType along with Values used in clauses (#2807)
The OpenMP clauses that take pointers will need the element type added
once opaque pointers is enabled. Since PointerType::getElementType()
can no longer be used, the element type needs to be passed around with
the llvm::Value. This change modifies DefinedValues and ReferencedValues
so it contains a Value/Type pair instead of just a Value. The tracking
logic isn't changed so the IR should be the same as before.

The only IR change is the addition of TYPED clauses for the address
version of num_teams and thread_limit. This is only visible when the
-fopenmp-typed-clauses option is used. See CMPLRLLVM-28637.