CMPLRLLVM-34307: Removed DenseHLNodeMapInfo to fix use-after-free issue (#2767)
DenseHLNodeMapInfo class for added in an attempt to form 'stable'
order of nodes in DenseMap. This was a bad idea in itself because
DenseMap is a hash table which can be dynamically resized. The order
of the nodes can change with the resizing of the table as same nodes
can get different slots. Hash table isn't the right data structure
if one wants stable order of nodes.

The problem reported in the JR though was a use-after-free scenario.
It goes like this-

1) Framework stores a bunch of HLIfs in a DenseMap (HLIf to bblock
   mapping).

2) Some of the HLIfs are optimized away by loop formation phase in the
   framework (elimination of loop bottom test) but the entries remain in
   the DenseMap.

3) Some new HLIfs are formed by the framework to clone ZTTs. These are
   added to the DenseMap. DenseMap decides to grow its size and calls
   HLIf->getNumber() to recompute hash function for existing entries in
   the map. This hash function getHashValue() is specified by
   DenseHLNodeMapInfo. Since some of the HLIfs have been freed, the hash
   function is accessing free'd memory.