CMPLRLLVM-34998: Update UTF BOM test cases (#2863)
This change set updates the lit tests for checking if
the byte order marks in UTF where handled correctly. Rather
than use 'grep' to check if the input file is in the correct
encoding, we are going to supply a simple python script that
opens the file with the input encoding and checks if the file
was read correctly.