CMPLRLLVM-34154: Support for UTF-8 BOM files when reading COFF module-definition files (#2820)
* CMPLRLLVM-34154: Support for UTF-8 BOM files when reading COFF module-definition files

This change set adds the support for reading UTF-8 byte order
mark (BOM) COFF module-definition (.def) files. The issue is
that BOM files have a number of leading special characters
before the actual text, compared to the non-BOM files. In
the case of UTF-8, there are 3 leading special characters.
This change set checks for the leading characters after
reading the file, remove them from the string, and pass
the new string to the parser to collect the information
from the file.

These are the same changes as PR #2460, but for xmain-rel.

* Fix for UTF test cases (changes are in xmain PR #2863)