Lower complex intrinsics correctly for the Windows ABI. (#2733)
MSVC itself doesn't support _Complex types directly, but clang-cl lowers these
types as if they were struct return values. This means that _Complex float is
returned via edx:eax, as in the SysV ABI (even in 64-bit), which necessitates
the support for this mode. Although that didn't turn out to be as annoying as I
thought to implement originally.

(Note that libirc implements x86_fp80 support, even though long double is mapped
into double, not x86_fp80, in the Windows ABI).