[DTrans - OpaquePtr] Propagate DTrans type metadata onto allocas during SROA (#2817)
When DTrans type metadata is being used for performing the DTrans analysis,
information needs to be propagated onto the alloca instructions produced
by SROA in some cases. This handles setting the DTrans metadata on the
alloca when the value type of the alloca contains a pointer type, and the
alloca was formed by taking one or more fields from an alloca that previously
was creating a structure type.