[DPCPPKernelTransforms] Correct scalar/vectorized kernel name in metadata (#2938)
Since the function name is added with suffix in barrier pass, the metadata for
vectorized, masked vectorized and scalar kernel are set with incorrect function
name with suffix. That will lead to the useless function declaration with name
suffix not being removed and lead to PrepareKernelArgs pass failing.

Fixed by replacing it with original function name in the metadata for vectorized
kernel, masked kernel and scalar kernel in barrier pass.