[DTrans] Sort structures in output of DTransImmutableInfo debug trace (#3055)
While testing the collection of values for the DTransImmutableInfo pass
on 538.imagick to compare the result from the new DTrans safety analyzer
with the original DTrans analysis results, I found the ordering of the
output was arbitrary. Update the output under the -drans-print-immutable-types
trace to print with the structures sorted by name. No new LIT tests added
because the -dtrans-print-immutable-types option is already covered by
existing LIT tests.