[XMAIN-WEB][LLD]CMPLRLLVM-35249: Pass the real symbol's name when resolving a symbol (#3176)
The community recently removed adding the symbol name when creating a
temporary symbol to compare against another symbol. This temporary symbol
used to have the same name as the other symbol, but now the name will be
an empty string. The problem is that we use the symbol's name to find if
the it has a GNU linkonce section.

This change set basically extends the function resolve, and the called
helper functions, to pass the real symbol name when it is needed. The
function that compartes the GNU linkonce section will check if the
symbol has a name, else collect the passed real name and check for
the section.

The guilty commit from the community:
https://github.com/llvm/llvm-project/commit/977a1a523c2c510d847bb7814b1e29774fc02472