[DTRANS] CMPLRLLVM-32501: Fix the conditions to set base-padded structures in DTrans (Part 1) (#3132)
This change set updates the conditions to set the relationship between
base and padded structures in DTrans. Previously, we merged the
safety data when we found that two structures were base and padded.
In order to the merge, we needed to check for some safety violations
(e.g. BadCasting, mismatched element access, etc.). This prevented to
merge the information in some cases.

This change set proposes to use 2 safety bits rather than merging the
safety data. The structures with ABI passing will be set as
StructCouldHaveABIPadding, and the base structures will be marked
as StructCouldHaveABIPadding. Then we disabled all the transformations
that could be affected by the base-padded relationship but it isn't
supported or implemented (if one structure is modified, the same
transformation should be applied to the other) by checking if the
structure is set with one of the safety bits.

This is a two parts change set. The first one (this change set) fixes
the DTrans analysis. The next change set will extend delete fields
to support base-padded structures.