CMPLRLLVM-35374: Put recursive progression cloning for 548.exchange2 under Advanced AVX2. (#3183)
* CMPLRLLVM-35374: Put recursive progression cloning for 548.exchange2 under
Advanced AVX2.

Notice that I introduced a flag to be used with some of the LIT tests called
-ip-rec-pro-cloning-skip-cpu-check.  I added this to the old tests that were
checking for the AVX512 to AVX2 conversion. There wasn't any way to pass in
-mattr=+avx2 on the command line without also getting +avx2 added to the
function attributes that I was trying to check. I could have rewritten the
attribute checking to check for duplicates, but I didn't want to make a big
change to the compiler code just to accommodate the LIT tests. There are
plenty of other LIT test that check whether the optimization is enabled or
disabled under AVX2, including some new ones in this change set.

* Fix copyrights