[VPlan][HIR] Implement pumping for SVML calls in HIR vector codegen. (#2677)
This patch introduces an optimization for vectorizing SVML calls,
when the call cannot be widened for a chosen VF, by pumping them
with a lower VF. This is achieved by breaking down the call
arguments to the selected lower VF context and using them to emit
multiple vector calls determined by pumping factor. These call
results are subsequently combined back to be used in the original
VF context. The implementation mimics its equivalent in LLVM-IR
vector codegen.

JIRA: CMPLRLLVM-32669