[DynClone] (CMPLRLLVM-35416): Generalize runtime checks to DynClone. (#3212)
* [DynClone] (CMPLRLLVM-35416): Generalize runtime checks to DynClone.

1. If application has more than 8 large constants, shrink fields
to 32-bit by checking the range from -2*31 -1 to 2*31 and disable
all other fancy things like bitfields and encoding/decoding.
2. Currently, 14 bits are used for shrunken fields and 2 bits are
used for bit field candidates. Vary the bitfield size to 0 or 1 or 2
depending on the values of bitfields. If there is no bitfield candidate,
use 16-bit for shrunken field. If bitfield candidate needs only 1 field.
15-bits are used for shrunken field.
3. Right now, 0 to 0x3ff6 range is used for shrunken field. 0x3ff6 to
0x3fff is used for decoding/encoding large constants. Change it vary this
number depending on number of large constants in application. At most
only 8 are allowed. If application has only 4, use 0 to 0x3ffb is used
for shrunken field.

* Fix Robert and Chris's comments.