CMPLRLLVM-35422: Reorganize the support for GNU LTO objects (#3300)
* CMPLRLLVM-35422: Reorganize the support for GNU LTO objects

The community removed the derived class ArchiveFile from the
base InputFile class. They now traverse through and archive,
pull the members and add all the files as Lazy objects. The
problem is that we used the ArchiveFile to catch when a member
in an archive was a GNU LTO object, or a nested archive.

This change set extends the new method of reading archives to
catch GNU LTO objects in an archive and/or nested archives.
When we have GNU LTO objects in an archive, we need to do a
partial linking because now all the archive's members are
treated as lazy objects. We need to make sure that the order
of symbols resultion is preserved.

Guilty commit: https://reviews.llvm.org/D119074

* Fix for Robert's comments

* Fix for Robert's comment