[VPlan][HIR] Centralize and extend setting of novectorize/nounroll markers  (#3253)
Changes in this patch use the canonical IV Phi and the loop bottom test to see if
we can determine the IV upper bound which can then be used to determine the loop
trip count. This information is used to generate inner DO loops during HIR vector
code generation.

When a loop is vectorized we mark it as novectorize to stop it from being
considered for vectorization again. This marking is extended to inner loops
inside a vectorized outer loop. The nounroll marker setting is also
extended to all top level vectorized loops. The code that does this is now
centralized in one place.