[XMAIN-WEB] CMPLRLLVM-35555: Skip symbols with GNU linkonce sections when looking for duplicate symbols in the post parsing phase (#3351)
* [XMAIN-WEB] CMPLRLLVM-35555: Skip symbols with GNU linkonce sections when looking for duplicate symbols in the post parsing phase

The community recently added a function to check for duplicated symbols
after reading all files. It happens twice: one before LTO and one after
LTO. The process doesn't understand the GNU linkonce sections because
the community deal with these sections using a hack.

This change set expands the check for duplicate symbols in the object
file post parsing phase to check for GNU linkonce sections. If the
symbol contains the information then it should be skept and the
resolution phase should be able to make a decision. Also, this change
disables the community test for GNU linkonce section since it was
created based on a hack. It adds 2 new test cases that checks for
the proper GNU linkonce section and duplicated symbols.

Guilty commit: https://reviews.llvm.org/D119908

* Fix for Robert' comments

* Fix for Chris and Robert comments