DTRANS/OpaquePTR: Stop depending on TypeFinder to find all types for MD (#3380)
As reported in CMPLRLLVM-35550, when we switch to opaque-pointers, the
type finder now no longer finds things that were reachable ONLY via
pointers. However, we still need metadata for those types.

Filtering the types in some way is necessary, as otherwise we end up
with a large number of types emitted to the Metadata, and since they are
referenced there, ALSO to the type-list at the top of the IR.

This patch uses the TypeFinder to find the 'root' types, then any types
that we end up reaching via those types. Fortunately, we don't have to
do anything special about things pointed to by a function parameter,
since the type-finder ends up finding in the already existing metadata.