[XMAIN-WEB][LLD] CMPLRLLVM-35637: Include -fPIC in the GNU LTO linking command (#3423)
* [XMAIN-WEB][LLD] CMPLRLLVM-35637: Include -fPIC in the GNU LTO linking command

The community recently introduced creating the local symbols
in parallel. The problem is that the changes tries to put
at most local symbols as possible. The issue is that when
creating the relocatable object from the GNU LTO linking,
it generates the symbols with relocation sections and this
is not allowed for local symbols. Including -fPIC in the
linking command solves the problem.

Guilty commit: https://reviews.llvm.org/D119909

* Fix for reading GNU LTO objects earlier when reading input files