[DTrans - OpaquePtr] Extend AOS-to-SOA allocation site uses to support 1 or 2 operand GEPs (#3435)
When opaque pointers are in use, the result of the allocation call for the structure
being transformed can be used directly in GEP instructions without an intervening
bitcast instruction. Users that access field addresses (2 index operands) or that
access the starting address of one of the array elements (1 index operands) need
to be supported. However, byte-flattened GEPs of the allocation should not be
supported.
[CMPLRLLVM-35617]