[CMPLRLLVM-35601] Update test cases for SplitIndirectBrCriticalEdges to work followiing pulldown (#3442)
The function SplitIndirectBrCriticalEdges contains INTEL_CUSTOMIZATIONS that cause
the IR output from it to differ from the community. The change in
https://reviews.llvm.org/D120096 caused some additional cases to have edges split,
which results in differences in xmain because the community relies on a cloned
block to be used as a target that will branch to the split block, but xmain directs
the edge directly to the split block. This causes differences in the hash for the
function used for PGO, as well as which blocks get counters associated with them.