CMPLRLLVM-35609: Restore inlining needed for SOA-to-AOS in new pass manager. (#3461)
* CMPLRLLVM-35609: Restore inlining needed for SOA-to-AOS in new pass manager.

Inlining differences between the old and new pass managers are inhibiting
SOA-to-AOS in the new pass manager. Prasad noted taht this is due to inlining
differences that inhibit indirect call specialization.  I found that a key
difference between the inlining in the old and new pass manager is that
the inliner in the new pass manager uses BlockFrequencyInfo (i.e. static
profiling) to guide the identification of cold call sites, which are subject
to a reduced inlining threshold (making it harder to inline them).

I have put in a change to InlineCost.cpp to not identify
intel-mempool-destructors as cold call sites. These are only identified as
such under qopt-mem-layout-trans. Prasad and I discussed, and I tried out,
additional conditions on the heuristics, but these proved to cost more
compile time without significantly reducing the number of call sites
identified.  So, I decided to go with a simple heuristic.

* Put this under SW_ADVANCED