CMPLRLLVM-35575: Guard the upcasting prevention in inst-combine when opaque pointers are used. (#3472)
The function that prevents the upcasting in inst-combine checks for
pointer element type. In the case of opaque pointers, this could be
a problem. There is a chance that if two instructions loads the
same address using different types, and one of the types is an
opaque pointer. In that case the function will try to collect a
pointer element type from an opaque pointer.

This change set basically guards the function to check if at least
one loaded instruction is an opaque pointer, then we are not going
to prevent the upcasting (load instruction simplification will run).