[VPlan][HIR] Don't invoke setBoundsForVectorLoop on inner loops. (#3465)
The utility targets outermost loops that are being vectorized.
Invoking it for inner loops leads to an assertion failure in
VPlan for unfound normalized loop IV. This patch fixes this bug
in merged CFG-based CG in HIR vectorizer. Additionally an affected
LIT test is setup for a bug related to liveness of dropped external
values. This bug will be fixed in subsequent PR.