[DTRANS] CMPLRLLVM-35063: Extend the DTrans base-padded structures analysis to work with opaque pointers (Part 1) (#3533)
* [DTRANS] CMPLRLLVM-35063: Extend the DTrans base-padded structures analysis to work with opaque pointers (Part 1)

This is the first part for porting the data collection and analysis
for the base-padded structures in DTrans for opaque pointers. The
implementation is basically the same as the typed pointers version,
with the exception that the related types map use DTransType rather
than llvm::Type. The rest is pretty straight forward.

There are 2 issues that still needs to be solved:

  1) If the memory is allocated with base type, and then accessed
     as padded type currently produces a Bad Casting when it shouldn't.

  2) If memcpy is used between the base and padded type is producing
     Bad memfunc, when it shouldn't.

* Fix for Robert' comments and possible issue with fields that are pointers

* Fix for Chris' comments

* Fix for undeclared DTransTestPaddedStructs cl option