[CMPLRLLVM-35703] [NPM] Generalize @llvm.assume intrinsic analysis in DTrans WeakAlign (#3554)
* [CMPLRLLVM-35703] [NPM] Generalize @llvm.assume intrinsic analysis in DTrans WeakAlign

Extend the analysis done for @llvm.assume by the DTrans WeakAlign transformation
to allow uses that do not appear to be related to pointers. When the new pass
manager is run, there are additional uses of this intrinsic being seen when
analyzing the IR for the weak align transformation. This PR generalizes support
of cases that are permitted to be used.

* Update comments

* Fix typos

* Fix LIT test run lines