Added support for NumTeams in VPOParoptConfig. (#3155)
Since fixed number of teams cannot be used with specific ND-range,
we have to disable specific ND-range partitioning and collapsing,
when VPOParoptConfig specifies number of teams for a target region.

I guess the cleanest way to do that is to add NUM_TEAMS and THREAD_LIMIT
clauses based on VPOParoptConfig. This has to be done early, i.e. before
any pass that may behave differently based on the presence of these clauses.
So I added a new pass that translates VPOParoptConfig into the clauses
before the loop collapsing pass. Now the clauses are added both
for the host and device compilations, so the passes behavior should be
aligned between the two compilations pretty well.

This changed pass pipelines quite a bit, and llvm/test/Other LIT tests
have to be carefully reviewed for proper INTEL_CUSTOMIZATION marks,
though, I am not sure that we can do it cleanly (given that the early
VPOParoptApplyConfig pass changes the analysis pipeline a lot).

Note that the SIMD width is still translated from VPOParoptConfig
during Paropt transform. This may be re-worked, if we add a special
clause for TARGET region that will define the requested SIMD width.

Co-authored-by: Abhinav Gaba <abhinav.gaba@intel.com>