Fix allow_cpu_feature value checking, special case the 'generic' (#3550)
We were incorrectly checking 'page 2' of the values against the 'page 1'
list of valid values, which resulted in us rejecting page 2's value of
'1'.  Additionally, it allowed a bunch of others that are not yet
populated in page 2 with a '+' (where it is typically +featname).

This patch also special cases the 'generic' value ('1' in page 1), since
this is currently exposed in the library.  There is no corresponding
feature, so this shouldn't change the IR, but it shouldn't be an error
either.