Fix incorrect phi handling in exit blocks after cloning loop (#3452)
* Fix incorrect phi handling in exit blocks after cloning loop

After cloning loop, the predecessors of exit blocks are doubled. Each
incoming basicblocks of the original phinode in exit blocks has a
counterpart in clone loop. If the incoming value is def inside loop,
its counterpart is the def inside clone loop. Otherwise it is the def
outside any clone loop. Therefore we should add each counterparts of
imcomings of original loop to phinode in exit blocks.