[NFC][VPlan][HIR] Add test to reveal a liveness related bug for new VPExternalUses in merged CFG-based CG. (#3547)
VPlan's live in/out list creator transform creates new VPExternalUse
for some cases while creating in/outs for reductions (example -
reduction not used outside loop, in-memory reduction, etc.). This
leads to an existing DDRef being made newly liveout from the
original HLLoop but it is not added in the HLLoop liveout list. This
patch adds a test to reveal this bug which is then addressed in next PR.