Revert "CMPLRLLVM-35219: preserve use of lld for compiler-rt testing" (#3526)
This reverts commit 2f6d9b449238429db8d1ba376e69db4e035e4cfd.

This PR effectively switched compiler-rt testing to use lld on Windows, matching the xmain configuration so that we don't produce invalid binaries in combination with `-gcodeview-ghash`. This is no longer needed because we have disabled `-gcodeview-ghash` in xmain, and no longer desirable because it turns out that compiler-rt testing has some failures specific to lld. It also has failures with link.exe, but those are already captured so we'll go with those. See CMPLRLLVM-35688.