CMPLRLLVM-35479: Initial opaque pointer support for DTrans inlining. (#3617)
This provides the initial opaque pointer support for DTrans inlining and
non-inlining. MEMMANAGETRANS still needs to be completed by Prasad before this
can be finished. (Note the TODO in DTransForceInlineOP.cpp.)

To implement this support , I introduced a new pass that will set the
attributes to force or inhibit inlining for DTrans. Doing this allows all of
the relevant complicated code to reside in the DTrans Transforms directory,
and avoids excessive declarations of inline functions in header files. Now,
code that was in SOAToAOSExternal.h appears in SOATOSOAInternal.[h,cpp].

The changes include a bug fix for the old pass manager pipeline. Some of the
inlining setup code was not happening early enough. In the case where profiling
is present, the inliner could run before all of the inlining setup passes.
The soatoaos-inlpgo.ll LIT test tests for this condition.