Support new encoding (#3330)
// INTEL_FEATURE_XUCC
This patch is to support XuCC encoding and decoding.

Adding one new triple x86_64_xucc. It work similar to how 16/32/64-bit mode works.
Adding three new prefix: XuCCPD, XuCCXS and XuCCXD. The encoding of these prefix is same with mandatory prefix.
Only the XuCCPD can be used alone. In that time, it won't be a operand-size prefix. For XuCCXS and XuCCXD, it must have PD prefix.
// end INTEL_FEATURE_XUCC