[LIT] Fix detached thread in USMTest/memBlockingFree (#322)
There is a redundant std::thread creation and detach in memBlockingFree
test. The second kernel launch won't wait for the thread to finish,
therefore, there could be a situation that the test process terminates
but the second std::thread hasn't finished yet. This is dangerous.

This patch removes the use of std::thread which complicates multi-
threaded OpenCL kernel execution. Simply increasing global size can also
enables testing of clMemBlockingFreeINTEL.

This patch also splits the test so that in the future it is easier to
identify which part of the test causes fail.