[DTrans - OpaqwuePtr] Fix AOS-to-SOA conversion of Alloca instruction under opaque pointers (#3693)
Alloca instruction allocating a pointer to the structure being transformed need to be
changed to allocate an integer of the peeling index type. This was occuring with typed
pointers, but not when opaque pointer were used. Also, add safety check to disallow
AOS-to-SOA when there is an array of pointers to the type being transformed because the
transformation code does not handle converting users of array.

[CMPLRLLVM-36039]