[VPlan][HIR] Setup liveness info for external VPValues flushed during merged CFG-based CG. (#3496)
* [VPlan][HIR] Setup liveness info for external VPValues flushed during merged CFG-based CG.

HIR vectorizer's merged CFG-based CG flushes VPExternalDefs and
VPConstants that are mapped in VPValWideRef for a specific VF
since the same VPValue can be used in context of multiple VFs.
This leads to loss of liveness info for DDRefs that are created
while widening these VPValues. This patch addresses the above
issue by updating setupLiveInLiveOut utility to visit VPValues
that are flushed during CG. We additionally start recording
DDRefs that are created for widening these VPValues.